#151 move tests to a subpackage
