+++ src/test/java/org/mapstruct/intellij/inspection/inheritance/AutoInheritConfigurationTest.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache License version 2.0, available at https://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.intellij.bugs._148;
+package org.mapstruct.intellij.inspection.inheritance;
 
 import org.jetbrains.annotations.NotNull;
 import org.mapstruct.intellij.inspection.BaseInspectionTest;
@@ -16,7 +16,7 @@ public class AutoInheritConfigurationTest extends BaseInspectionTest {
 
     @Override
     protected String getTestDataPath() {
-        return "testData/bugs/_148";
+        return "testData/inspection/inheritance/auto";
     }
 
     @NotNull
