+++ src/test/java/org/mapstruct/intellij/inspection/inheritance/InheritConfigurationInspectionTest.java
@@ -4,18 +4,27 @@
  * Licensed under the Apache License version 2.0, available at https://www.apache.org/licenses/LICENSE-2.0
  */
 
-package org.mapstruct.intellij.inspection;
+package org.mapstruct.intellij.inspection.inheritance;
 
 import java.util.List;
 
 import com.intellij.codeInsight.daemon.impl.HighlightInfo;
 import com.intellij.lang.annotation.HighlightSeverity;
-import com.intellij.pom.java.LanguageLevel;
 import org.jetbrains.annotations.NotNull;
+import org.mapstruct.intellij.inspection.BaseInspectionTest;
+import org.mapstruct.intellij.inspection.UnmappedTargetPropertiesInspection;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+/**
+ * @author Oliver Erhart
+ */
 public class InheritConfigurationInspectionTest extends BaseInspectionTest {
+
+    @Override
+    protected String getTestDataPath() {
+        return "testData/inspection/inheritance";
+    }
 
     @NotNull
     @Override
@@ -24,15 +33,10 @@ public class InheritConfigurationInspectionTest extends BaseInspectionTest {
     }
 
     @Override
-    protected LanguageLevel getLanguageLevel() {
-        return super.getLanguageLevel();
-    }
-
-    @Override
     protected void setUp() throws Exception {
         super.setUp();
 
-        addDirectoryToProject( "../mapping/dto" );
+        addDirectoryToProject( "../../mapping/dto" );
     }
 
     public void testInheritConfigurationByInheritanceMapper() {
