+++ testData/inspection/inheritance/InheritConfigurationByInheritanceMapper.java
