+++ testData/inspection/inheritance/InheritConfigurationByMapperConfigMapper.java
