+++ testData/inspection/inheritance/InheritConfigurationByNameMapper.java
