+++ testData/inspection/inheritance/InheritConfigurationBySuperMapperMapper.java
