+++ testData/inspection/inheritance/InheritConfigurationInSameClassMapper.java
