+++ testData/inspection/inheritance/InheritConfigurationInSameClassOfUpdateMethodMapper.java
