+++ testData/inspection/inheritance/InheritConfigurationInSameClassWithContextMapper.java
