+++ testData/inspection/inheritance/InheritConfigurationNotInheritedByUsedMapper.java
