+++ testData/inspection/inheritance/auto/InheritanceStrategyAutoInheritAllFromConfigForward.java
