+++ testData/inspection/inheritance/auto/InheritanceStrategyAutoInheritAllFromConfigForwardNotConfigured.java
