+++ testData/inspection/inheritance/auto/InheritanceStrategyAutoInheritAllFromConfigReverse.java
