+++ testData/inspection/inheritance/auto/InheritanceStrategyAutoInheritAllFromConfigReverseNotConfigured.java
