+++ testData/inspection/inheritance/auto/InheritanceStrategyAutoInheritFromConfig.java
