+++ testData/inspection/inheritance/auto/InheritanceStrategyAutoInheritFromConfigNotConfigured.java
