+++ testData/inspection/inheritance/auto/InheritanceStrategyAutoInheritReverseFromConfig.java
