+++ testData/inspection/inheritance/auto/InheritanceStrategyAutoInheritReverseFromConfigNotConfigured.java
