+++ testData/inspection/inheritance/auto/InheritanceStrategyExplicitAnnotated.java
