+++ testData/inspection/inheritance/auto/InheritanceStrategyExplicitNotAnnotated.java
